Updates Edificio

+++ oo2/src/main/java/sistema/integrador/oo2/services/IEdificioService.java
@@ -8,7 +8,7 @@ import sistema.integrador.oo2.entities.Edificio;
 public interface IEdificioService {
 	public List<Edificio>listar();
 	public Edificio guardarEdificio(Edificio edificio);
-	public Edificio obtenerEdificio(Long id);
+	public Edificio obtenerEdificio(int id);
 	public Edificio actualizarEdificio(Edificio edificio);
-	public void eliminarEdificio(Long id);
+	public void eliminarEdificio(int id);
 }
